refactor(Personagens): alterar vida para double

+++ src/modulo-01-java-OO/Modulo01/Dwarf.java
@@ -25,7 +25,7 @@ public class Dwarf extends Personagem {
             this.experiencia += 2;
         }
         else if (numero > 100) {
-            int dano = 10, vidaAposFlechada = this.vida - dano;
+            double dano = 10, vidaAposFlechada = this.vida - dano;
             if (vidaAposFlechada == 0) {
                 this.status = Status.MORTO;
             }
